Update test-related and cmake and scripts

+++ jenkins/run_tests.sh
@@ -1,25 +1,60 @@
 #!/bin/bash
 
+if [ "$WORKSPACE" == "" ] ; then
+  WORKSPACE=$PWD
+fi
+
+if [ "$COMPILER" == "" ] ; then
+  COMPILER=g++
+fi
+
+if [ "$CONCURRENCY" == "" ] ; then
+  CONCURRENCY=5
+fi
+
+echo "WORKSPACE:   ${WORKSPACE}"
+echo "COMPILER:    ${COMPILER}"
+echo "CONCURRENCY: ${CONCURRENCY}"
+
 cd ${WORKSPACE}/build/debug
 
-echo "cleaning previous compilation"
-rm -rf * >& /dev/null
+if [ -z ${NOBUILD} ] ; then
 
-../build.sh g++
+  echo "cleaning previous compilation"
+  rm -rf * >& /dev/null
 
-if [ $? -ne 0 ]; then
-  echo "cmake failed"
-  exit 1
+  ../build.sh ${COMPILER}
+
+  if [ $? -ne 0 ]; then
+    # exit immediately if cmake fails
+    echo "$module: cmake failed"
+    exit 1
+  fi
+
+  make -j${CONCURRENCY}
+
+  if [ $? -ne 0 ] ; then
+    # exit immediately if make fails
+    echo "$module: make failed"
+    exit 1
+  fi
+
 fi
 
-make -j5
+ret=0
+for testunit in $(find . | grep 'Test/test') ; do
 
-if [ $? -ne 0 ] ; then
-  echo "make failed"
-fi
+  report=$(basename $testunit)
 
-Test/testmain \
-  --gtest_shuffle \
-  --gtest_output=xml:${WORKSPACE}/testunit.xml
+  $testunit \
+    --gtest_shuffle \
+    --gtest_output=xml:${WORKSPACE}/$report.xml
 
-exit $?
+  if [ $? -ne 0 ] ; then
+    # if a test unit fails, mark the whole execution as failed but go on
+    ret=1
+  fi
+
+done
+
+exit $ret
